requirements.txt
requests==2.31.0
pandas==2.2.1
python-dotenv==1.0.1
supabase==2.3.5
streamlit==1.31.1
st-supabase-connection==1.1.0
plotly==5.19.0
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3 